refactor: move 'clickedUI' to ui.js
