+++ src/events/raycast.js
@@ -3,7 +3,7 @@
 import * as THREE from 'three';
 import * as TRANSFORM from '/src/controls/transform.js';
 import {scene, camera} from '/src/main.js';
-import UI from '/src/configs/ui.js';
+import {clickedUI} from '/src/utils/ui.js';
 import {raycasterObjects} from '/src/utils/object.js';
 
 const pointer = new THREE.Vector2();
@@ -18,27 +18,6 @@ window.addEventListener('mousemove', event => {
   pointer.x = (event.clientX / window.innerWidth) * 2 - 1;
   pointer.y = - (event.clientY / window.innerHeight) * 2 + 1;
 });
-
-/**
- * Checks to see if event pressed is the 'scene' UI element
- * @param {event} event
- * @returns true if scene section is in event's path array
- */
-function clickedUI(event) {
-  for (let i = 0; i < event.composedPath().length; i++) {
-    if (event.composedPath()[i] === UI.file ||
-      event.composedPath()[i] === UI.edit ||
-      event.composedPath()[i] === UI.add ||
-      event.composedPath()[i] === UI.help ||
-      event.composedPath()[i] === UI.settings ||
-      event.composedPath()[i] === UI.viewHelper ||
-      event.composedPath()[i] === UI.hotkeys) {
-      return true;
-    }
-  }
-
-  return false;
-}
 
 /**
  * Do something when mouse is pressed over mesh
